requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
phonenumbers==9.0.9
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
